feat: avoid err by uncoind triple from target-lexicon
